display supervisor on class list

+++ views/students.php
@@ -23,9 +23,11 @@ $studentsList = $students->findAllPaginated($offset, $resultsPerPage);
       </div>
       <button class="btn"><i class="ri-equalizer-line"></i></button>
       <button class="btn"><i class="ri-sort-desc"></i></button>
-      <button type="button" class="add-btn btn">
-        <i class="ri-add-line"></i>
-      </button>
+      <?php if ($role == "admin"): ?>
+        <button type="button" class="add-btn btn">
+          <i class="ri-add-line"></i>
+        </button>
+      <?php endif ?>
     </div>
   </div>
 
@@ -38,7 +40,9 @@ $studentsList = $students->findAllPaginated($offset, $resultsPerPage);
         <th>Field of study</th>
         <th>Phone</th>
         <th>Address</th>
-        <th>Actions</th>
+        <?php if ($role == "admin"): ?>
+          <th>Actions</th>
+        <?php endif ?>
       </tr>
     </thead>
     <tbody>
@@ -51,21 +55,23 @@ $studentsList = $students->findAllPaginated($offset, $resultsPerPage);
           <td><?= htmlspecialchars($student['email']); ?></td>
           <td><?= htmlspecialchars($student['phone']); ?></td>
           <td><?= htmlspecialchars($student['address']); ?></td>
-          <td class="row" id="actions">
-            <button
-              class="btn edit-btn"
-              data-id="<?= $student['id'] ?>"
-              data-name="<?= htmlspecialchars($student['surname']) ?> <?= htmlspecialchars($student['name']) ?>"
-              data-email="<?= htmlspecialchars($student['email']) ?>"
-              data-level="<?= htmlspecialchars($classe['classId']) ?>"
-              data-phone="<?= htmlspecialchars($student['phone']) ?>"
-              <?php $parent = $parents->findById($student['parentId']) ?>
-              data-parent-name="<?= htmlspecialchars($parent['name']) ?>"
-              data-parent-email="<?= htmlspecialchars($parent['email']) ?>"
-              data-parent-phone="<?= htmlspecialchars($parent['phone']) ?>"
-              data-address="<?= htmlspecialchars($student['address']) ?>"><i class="ri-edit-box-line"></i></button>
-            <a class="btn edit-btn" href="/delete_students?id=<?= $student['id'] ?>"><i class="ri-delete-bin-6-line"></i></a>
-          </td>
+          <?php if ($role == "admin"): ?>
+            <td class="row" id="actions">
+              <button
+                class="btn edit-btn"
+                data-id="<?= $student['id'] ?>"
+                data-name="<?= htmlspecialchars($student['surname']) ?> <?= htmlspecialchars($student['name']) ?>"
+                data-email="<?= htmlspecialchars($student['email']) ?>"
+                data-level="<?= htmlspecialchars($classe['classId']) ?>"
+                data-phone="<?= htmlspecialchars($student['phone']) ?>"
+                <?php $parent = $parents->findById($student['parentId']) ?>
+                data-parent-name="<?= htmlspecialchars($parent['name']) ?>"
+                data-parent-email="<?= htmlspecialchars($parent['email']) ?>"
+                data-parent-phone="<?= htmlspecialchars($parent['phone']) ?>"
+                data-address="<?= htmlspecialchars($student['address']) ?>"><i class="ri-edit-box-line"></i></button>
+              <a class="btn edit-btn" href="/delete_students?id=<?= $student['id'] ?>"><i class="ri-delete-bin-6-line"></i></a>
+            </td>
+          <?php endif ?>
         </tr>
       <?php endforeach; ?>
     </tbody>
